fastapi-backend/pyproject.toml
[project]
name = "pm-backend"
version = "1.0.0"
description = "AI Assisted Project and Knowledge Management System"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    # FastAPI and server
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Database (PostgreSQL async)
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "psycopg2-binary>=2.9.9",

    # Object Storage
    "minio>=7.2.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # File type detection (magic bytes)
    "python-magic-bin>=0.4.14",

    # Utilities
    "python-multipart>=0.0.17",
    "pydantic-settings>=2.4.0",
    "websockets>=13.0",

    # Database migrations
    "alembic>=1.14.0",

    # Real-time collaboration (5000 concurrent users)
    "redis[hiredis]>=5.0.0",
    "msgpack>=1.0.0",

    # Full-text search
    "meilisearch-python-sdk>=5.5",

    # Background job processing
    "arq>=0.26.0",

    # AI providers
    "openai>=1.30.0",
    "anthropic>=0.40.0",
    "cryptography>=42.0.0",

    # Vector embeddings
    "pgvector>=0.3.0",
    "tiktoken>=0.7.0",
    "defusedxml>=0.7.0",

    # Email (SMTP)
    "aiosmtplib>=3.0.0",

    # AI agent SQL access & Excel export
    "openpyxl>=3.1.0",
    "sqlglot>=25.0.0",

    # AI agent (LangGraph)
    "langgraph>=1.0.0,<2.0.0",
    "langgraph-checkpoint-postgres>=3.0.0,<4.0.0",
    "psycopg[binary]>=3.1.0,<4.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    "langchain-anthropic>=1.0.0,<2.0.0",
    "sse-starlette>=2.0.0",

    # PDF export (pure Python, no system deps)
    "fpdf2>=2.8.0,<3.0.0",

    # Document import (PDF/DOCX/PPTX conversion)
    "docling>=2.0.0,<3.0.0",

    # Spreadsheet extraction (calamine for Excel, chardet for encoding)
    "python-calamine>=0.6.0,<1.0.0",
    "chardet>=5.0.0,<6.0.0",

    # Visio extraction
    "vsdx>=0.5.0,<1.0.0",

    # Web tools (search + scrape)
    "duckduckgo-search>=7.0.0,<8.0.0",
    "trafilatura>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
# copilotkit is optional — frontend uses custom REST endpoints for AI chat
# Install with: uv sync --extra copilotkit
copilotkit = ["copilotkit>=0.1.35,<1.0.0"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.9.0",
]
load = [
    "locust>=2.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
